change to int

+++ src/BmiService.java
@@ -1,5 +1,5 @@
 public class BmiService {
-    public double calculate (double bodyMassKg, double heightMetre) {
+    public int calculate (double bodyMassKg, double heightMetre) {
         double index = bodyMassKg / (heightMetre * heightMetre);
         return (int)index;
     }
